feat: expand Windows build matrix to include additional environments (#4)
- Expand Windows build matrix to include additional environments.
- Update build workflow to support MinGW and enhance cross-platform compatibility.
- Add project dependencies and additional source files for libfasteval and microgettext.
- Add centralized MSBuild configuration with Directory.Build.props and build-msbuild-all task.
- Include MSYS2 environment variable support and add WSL build task.

## improve compatibility and code clarity

 - Update environment variable names for clarity and consistency.
- Initialize pointer variables to NULL for safety.
- Add platform-specific compatibility for stat structures and type definitions.
- Suppress deprecated warnings for GetVersionEx in Windows compatibility code.
- Correct string comparison functions and Windows-specific argument handling.
- Add cross-platform file handling functions and vsnprintf wrapper.
- Simplify Makefiles by improving configuration and dependency handling.
- Address linker errors in pbatize and tea Makefiles with Windows-specific flags.
- Improve taskfile and .gitignore for better build and test management.

## enhance consistency and maintainability

- Rename and adjust type definitions to prevent conflicts (DIR and dirent).
- Reorder and simplify GitHub Actions workflows for clarity and efficiency.
- Update project filters and header settings for uniformity.
- Simplify handling of platform-specific macros and preprocessor directives.

## general project improvements

-  Update .gitignore to exclude build artifacts and directories.
- Remove unused configurations and redundant definitions.
- Improve licensing information in scanner.l.
- Add missing keywords and commands to dictionaries.

+++ libpbat/thread/pBat_Thread.c
@@ -177,7 +177,7 @@ LIBPBAT void     pBat_AbortThread(THREAD* lpThId)
 
 LIBPBAT void     pBat_EndThread(void* iReturn)
 {
-    ExitThread((DWORD)iReturn);
+    ExitThread((DWORD)(size_t)iReturn);
 }
 
 LIBPBAT int      pBat_WaitForThread(THREAD* thId, void** lpRet)
